Implement an internal test matrix
Molecule needs something at a higher level to decide what
scenarios should run and which actions are associated with
those scenarios.

This was somewhat tricky to do in a reasonable way.  Molecule maintains
a mapping of scnarios to actions to sequences.  This matrix is reported
at the beginning of a Molecule invocation, and will run the necessary
actions.  Unnecessary actions are filtered and reported as skipped.